// apps/cooltix-homework/graphql/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export type MakeEmpty<T extends { [key: string]: unknown }, K extends keyof T> = { [_ in K]?: never };
export type Incremental<T> = T | { [P in keyof T]?: P extends ' $fragmentName' | '__typename' ? T[P] : never };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type Member = {
  __typename?: 'Member';
  address: MemberAddress;
  email: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  lastName: Scalars['String']['output'];
  phoneNumber: Scalars['String']['output'];
  profilePictureUrl: Scalars['String']['output'];
};

export type MemberAddress = {
  __typename?: 'MemberAddress';
  addressLine: Scalars['String']['output'];
  city: Scalars['String']['output'];
  country: Scalars['String']['output'];
  postalCode: Scalars['String']['output'];
  state: Scalars['String']['output'];
};

export type Query = {
  __typename?: 'Query';
  allMembers: Array<Member>;
  member?: Maybe<Member>;
};


export type QueryMemberArgs = {
  id: Scalars['ID']['input'];
};

export type GetAllMembersQueryVariables = Exact<{ [key: string]: never; }>;


export type GetAllMembersQuery = { __typename?: 'Query', allMembers: Array<{ __typename?: 'Member', id: string, firstName: string, lastName: string, email: string, phoneNumber: string, profilePictureUrl: string, address: { __typename?: 'MemberAddress', country: string, state: string, postalCode: string, city: string, addressLine: string } }> };

export type GetMemberByIdQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type GetMemberByIdQuery = { __typename?: 'Query', member?: { __typename?: 'Member', id: string, firstName: string, lastName: string, email: string, phoneNumber: string, profilePictureUrl: string, address: { __typename?: 'MemberAddress', country: string, state: string, postalCode: string, city: string, addressLine: string } } | null };